+++ changelog.md
@@ -8,6 +8,7 @@
 **What's Fixed**
 * [useForm]: improved router block removal on discard and custom beforeLeave for close action. Rework useLock to unblock router immediately, rather than on next render
 * [Modals]: fixed incorrect order of abort() calls when pressing ESC with nested modals — now only the topmost modal responds to ESC key
+* Make clear buttons accessible via keyboards ([#2845](https://github.com/epam/UUI/pull/2845))
 
 # 6.2.0 - 05.08.2025
 **What's New**